new model

+++ train.py
@@ -20,19 +20,18 @@ device = torch.device("cuda:1" if torch.cuda.is_available() else "cpu")
 print(device)
 
 
-def train(train_dataset, val_dataset, model, epochs, lr, model_name, adversarial=False, checkpoint=True):
+def train(train_dataset, val_dataset, model, epochs, lr, model_name, adversarial=False, checkpoint=True, eps=10):
 
     print('Training {0} for {1} epochs'.format(model_name, epochs))
     # optimizer = torch.optim.Adam(model.parameters(), lr=lr, betas=(0.9, 0.999), weight_decay=5e-4)
     optimizer = torch.optim.SGD(model.parameters(), lr=lr, momentum=0.9, weight_decay=5e-4)
-    scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, milestones=[150, 250, 350], gamma=0.1)
+    scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, milestones=[90, 130, 350], gamma=0.1)
 
-    loss =  nn.CrossEntropyLoss() # cross entropy
+    loss =  nn.CrossEntropyLoss()
     plotted_train_epochs = []
     train_accuracies = []
     test_accuracies = []
     train_losses = []
-    # pgd_attack = torchattacks.PGD(model, eps=0.5, alpha=0.1, iters=5)
 
     for epoch in range(epochs):
         total_loss = 0
@@ -42,10 +41,8 @@ def train(train_dataset, val_dataset, model, epochs, lr, model_name, adversarial
             inputs, labels = batch[0].to(device), batch[1].to(device)
             total_examples_seen += len(inputs)
             if adversarial:
-                # inputs = pgd_attack(inputs, labels).to(device)
-                eps=10
+                # eps=20
                 inputs = pgd(inputs, labels, model, stepsize=eps*2.5/7, eps=eps, steps=7, constraint='l_2')
-                # inputs = fgsm(inputs, labels, model, eps=.08)
                 optimizer.zero_grad()
                 out = model(inputs)
             else:
@@ -87,20 +84,12 @@ def train(train_dataset, val_dataset, model, epochs, lr, model_name, adversarial
             torch.save(model.state_dict(), root + str(epoch))
 
     # done training
-    torch.save(model.state_dict(), root + 'final')
+    if checkpoint:
+        root = 'models/' + model_name + '/'
+        if not os.path.exists(root):
+            os.mkdir(root)
+            torch.save(model.state_dict(), root + 'final')
 
-
-
-# transform = transforms.Compose([transforms.RandomAffine(degrees=15, translate=(0.1, 0.1)),
-#                                 transforms.RandomHorizontalFlip(p=0.5),
-#                                 transforms.ToTensor(),
-#                                 transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
-# ])
-# train_transform = transforms.Compose([transforms.RandomAffine(degrees=15, translate=(0.1, 0.1)),
-#                                 transforms.RandomHorizontalFlip(p=0.5),
-#                                 transforms.ToTensor(),
-#                                 transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010))
-# ])
 train_transform = transforms.Compose([
     transforms.RandomHorizontalFlip(),
     transforms.RandomCrop(32, 4),
@@ -108,13 +97,11 @@ train_transform = transforms.Compose([
     transforms.ToTensor(),
     transforms.Normalize(mean=[0.485, 0.456, 0.406],
                                      std=[0.229, 0.224, 0.225]),
-                                     # , transforms.Lambda( lambda x: x / 255.)
 ])
 
 test_transform = transforms.Compose([transforms.ToTensor(),
                 transforms.Normalize(mean=[0.485, 0.456, 0.406],
                                      std=[0.229, 0.224, 0.225])
-                                     # , transforms.Lambda( lambda x: x / 255.)
 ])
 
 train_data = torchvision.datasets.CIFAR10('data/', download=True, train=True, transform=train_transform)
@@ -134,5 +121,5 @@ model.to(device)
 model.train()
 # print(model)
 # train(train_dataloader, test_dataloader, model, 100, 0.001, adversarial=False)
-train(train_dataloader, test_dataloader, model, 100, 0.1, adversarial=True, model_name='resnet50_l2eps=10')
+train(train_dataloader, test_dataloader, model, 150, 0.1, adversarial=True, model_name='resnet50_l2eps=100', eps=100)
 # train(train_dataloader, test_dataloader, model, 25, 0.1, adversarial=False, model_name='resnet18_normal')
